+++ app/src/main/java/com/example/ratemate/ui/screens/CurrencyDataScreen.kt
@@ -8,7 +8,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 
 @Composable
-fun CurrencyDataScreen() {
+fun CurrencyDataScreen(currencyCode: String) {
     Box(
         modifier = Modifier.fillMaxSize(),
         contentAlignment = Alignment.Center
